Struct vptree::vptree::VPTree

pub struct VPTree<F: Float, T: MetricItem<F>> {
    // some fields omitted
}

Vantage Point Tree

A vantage-point tree stores a set of points to be later queried against.

Methods

impl<F: Float, T: MetricItem<F>> VPTree<F, T>

fn new(items: Vec<T>) -> Option<VPTree<F, T>>

Construct a new vantage point tree from a set of elements.

Returns None if items is an empty vector.

new makes no effort to check that the MetricItem trait implementation actually defines a matric. If the metric is not defined correctly, the resulting tree may not yield correct answers for later queries.

fn within_radius<B: Borrow<T>>(&self, query: B, radius: F, sorted: bool) -> Vec<&T>

Return all elements with a given radius of the target.

If sorted is true, the elements are sorted by ascending distance from the query point,

fn nearest_neighbor<B: Borrow<T>>(&self, query: B) -> &T

Find the closest point in tree to query.

fn nearest_neighbors<B: Borrow<T>>(&self, query: B, k: usize, sorted: bool) -> Vec<&T>

Find the k points in the tree closest to query.

If sorted is true, the returned points will be sorted by distance to query. If k is larger than the total number of points in the tree, all of the points in the tree are returned.

impl<F: Float + Display, T: MetricItem<F> + Debug> VPTree<F, T>

fn dump(&self) -> String

Return a pretty-printed recursive description of the entire tree.

This function is mainly intended for debugging.