Trait vptree::vptree::MetricItem

pub trait MetricItem<F: Float> {
    fn distance(&self, b: &Self) -> F;
}

Defines a metric for items in a metric space.

A metric is a function on a set S, with the following properties.

f(x, y) >= 0, with equality iff x == y

f(x, y) = f(y, x)

f(x, y) + f(y, z) <= f(x, z)

A VP-Tree can only be constructed by a set forming a metric. If the distance function does not satisfy the metric conditions, a vp-tree constructed from the elements will not be correct.

Required Methods

fn distance(&self, b: &Self) -> F

Return the distance to another element in the metric space.

The distance function must satisfy the metric properties.